requirements.txt
astroid==1.5.3
backports.functools-lru-cache==1.4
configparser==3.5.0
Django==1.11.5
djangorestframework==3.6.4
enum34==1.1.6
freezegun==0.3.9
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pep8==1.7.0
PyJWT==1.5.3
pylint==1.7.2
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
python-dateutil==2.6.1
pytz==2017.2
singledispatch==3.4.0.3
six==1.11.0
wrapt==1.10.11
